Proper Gimble Spin - Head rotation
Added correct code to properly spin the dome head using the magnets and a 360 degree turn servo in the gimble.  If you use a 180 degree servo you will need to change lines: 505 and 531.  Although 531 is remarked.
This should correctly stop the servo from spinning when channel 5 which is the combination of R2 of Left Joy + Left and Right of Right Joy is pressed...